requirements.txt
Django==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.1
googlemaps==3.0.2
numpy==1.16.0
python-decouple==3.1
pytz==2018.9
scipy==1.2.0
